backend/pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "backend"
version = "1.0.0"
description = "Al Imran Fabrics - Multi-brand Women's Fashion E-commerce Platform"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "sqlalchemy>=2.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.6",
    "aiosqlite>=0.19.0",
    "pandas>=2.0.0",
    "alembic>=1.12.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.3.0",
    "email-validator>=2.0.0",
]

[tool.setuptools]
packages = ["app"]
